pyproject.toml
[project]
name = "rag-langchain"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Neuronspeeed", email = "andrei.moise@riskeyeque.com" }
]
dependencies = [
    "langchain-community>=0.3.0",
    "ipykernel>=6.29.5",
    "ipynb>=0.5.1",
    "jupyterlab-widgets>=3.0.13",
    "jupyterlab-pygments>=0.3.0",
    "l>=0.11.0",
    "jupyter-server>=2.14.2",
    "jupyterlab>=4.2.5",
    "jupyter>=1.1.1",
    "langchain-openai>=0.2.0",
    "gitpython>=3.1.43",
    "langgraph>=0.2.22",
    "pygraphviz>=1.13",
    "matplotlib>=3.9.2",
    "plotly>=5.24.1",
    "networkx>=3.3",
    "chromadb>=0.5.7",
    "loadotenv>=1.0.1",
    "langchain>=0.3.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.1",
    "langchain-chroma>=0.1.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rag_langchain"]
